fix(logcollector): add tags property to CollectingNode
