+++ sdcm/utils/common.py
@@ -49,6 +49,7 @@ import libcloud.storage.types
 from libcloud.compute.providers import get_driver
 from libcloud.compute.types import Provider
 
+from sdcm.remote import LOCALRUNNER
 from sdcm.utils.ssh_agent import SSHAgent
 from sdcm.utils.decorators import retrying
 
@@ -1562,3 +1563,34 @@ def normalize_ipv6_url(ip_address):
     if ":" in ip_address:  # IPv6
         return "[%s]" % ip_address
     return ip_address
+
+
+def get_username():
+
+    def is_email_in_scylladb_domain(email_addr):
+        return bool(email_addr and "@scylladb.com" in email_addr)
+
+    def get_email_user(email_addr):
+        return email_addr.strip().split("@")[0]
+
+    # First check that we running on Jenkins try to get user email
+    email = os.environ.get('BUILD_USER_EMAIL')
+    if is_email_in_scylladb_domain(email):
+        return get_email_user(email)
+    user_id = os.environ.get('BUILD_USER_ID')
+    if user_id:
+        return user_id
+    current_linux_user = getpass.getuser()
+    if current_linux_user == "jenkins":
+        return current_linux_user
+    # We are not on Jenkins and running in Hydra, try to get email from Git
+    # when running in Hydra there are env issues so we pass it using SCT_GIT_USER_EMAIL variable before docker run
+    git_user_email = os.environ.get('GIT_USER_EMAIL')
+    if is_email_in_scylladb_domain(git_user_email):
+        return get_email_user(git_user_email)
+    # we are outside of Hydra
+    res = LOCALRUNNER.run(cmd="git config --get user.email", ignore_status=True)
+    if is_email_in_scylladb_domain(res.stdout):
+        return get_email_user(res.stdout)
+    # we didn't find email, fallback to current user with unknown email user identifier
+    return "linux_user={}".format(current_linux_user)
